% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{get_metrics}
\alias{get_metrics}
\alias{metrics}
\title{Get Metrics}
\usage{
metrics(actual, predicted)
}
\arguments{
\item{actual}{Actual values}

\item{predicted}{Predicted values}
}
\value{
List of confusion matrix, balanced accuracy, chance level, and binomial test
}
\description{
Get metrics and confusion matrix
}
