% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordtools-package.R
\docType{package}
\name{ordtools-package}
\alias{ordtools}
\alias{ordtools-package}
\title{ordtools: Ordinal regression tools}
\description{
Helper functions for ordinal regression
}
\author{
\strong{Maintainer}: Lincoln Colling \email{lincoln@colling.net.net}

}
\keyword{internal}
