% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.r
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(data, outcome, size)
}
\arguments{
\item{data}{Dataframe}

\item{outcome}{Outcome variable}

\item{size}{Number of observations in test data}
}
\value{
List of training and test data
}
\description{
Split data into training and test data
}
